--- src/assets/scss/items/study/study.scss
@@ -1,2 +1,0 @@
-@import "../../variable";
-
