+++ src/assets/scss/layout/footer.css
@@ -513,244 +513,244 @@ h5 {
   margin: 100px; }
 
 .margin-top-10 {
-  margin-top: 5px 0px 0px 0px; }
+  margin-top: 5px; }
 
 .margin-top-20 {
-  margin-top: 10px 0px 0px 0px; }
+  margin-top: 10px; }
 
 .margin-top-30 {
-  margin-top: 15px 0px 0px 0px; }
+  margin-top: 15px; }
 
 .margin-top-40 {
-  margin-top: 20px 0px 0px 0px; }
+  margin-top: 20px; }
 
 .margin-top-50 {
-  margin-top: 25px 0px 0px 0px; }
+  margin-top: 25px; }
 
 .margin-top-60 {
-  margin-top: 30px 0px 0px 0px; }
+  margin-top: 30px; }
 
 .margin-top-70 {
-  margin-top: 35px 0px 0px 0px; }
+  margin-top: 35px; }
 
 .margin-top-80 {
-  margin-top: 40px 0px 0px 0px; }
+  margin-top: 40px; }
 
 .margin-top-90 {
-  margin-top: 45px 0px 0px 0px; }
+  margin-top: 45px; }
 
 .margin-top-100 {
-  margin-top: 50px 0px 0px 0px; }
+  margin-top: 50px; }
 
 .margin-top-110 {
-  margin-top: 55px 0px 0px 0px; }
+  margin-top: 55px; }
 
 .margin-top-120 {
-  margin-top: 60px 0px 0px 0px; }
+  margin-top: 60px; }
 
 .margin-top-130 {
-  margin-top: 65px 0px 0px 0px; }
+  margin-top: 65px; }
 
 .margin-top-140 {
-  margin-top: 70px 0px 0px 0px; }
+  margin-top: 70px; }
 
 .margin-top-150 {
-  margin-top: 75px 0px 0px 0px; }
+  margin-top: 75px; }
 
 .margin-top-160 {
-  margin-top: 80px 0px 0px 0px; }
+  margin-top: 80px; }
 
 .margin-top-170 {
-  margin-top: 85px 0px 0px 0px; }
+  margin-top: 85px; }
 
 .margin-top-180 {
-  margin-top: 90px 0px 0px 0px; }
+  margin-top: 90px; }
 
 .margin-top-190 {
-  margin-top: 95px 0px 0px 0px; }
+  margin-top: 95px; }
 
 .margin-top-200 {
-  margin-top: 100px 0px 0px 0px; }
+  margin-top: 100px; }
 
 .margin-bottom-10 {
-  margin-bottom: 0px 0px 5px 0px; }
+  margin-bottom: 5px; }
 
 .margin-bottom-20 {
-  margin-bottom: 0px 0px 10px 0px; }
+  margin-bottom: 10px; }
 
 .margin-bottom-30 {
-  margin-bottom: 0px 0px 15px 0px; }
+  margin-bottom: 15px; }
 
 .margin-bottom-40 {
-  margin-bottom: 0px 0px 20px 0px; }
+  margin-bottom: 20px; }
 
 .margin-bottom-50 {
-  margin-bottom: 0px 0px 25px 0px; }
+  margin-bottom: 25px; }
 
 .margin-bottom-60 {
-  margin-bottom: 0px 0px 30px 0px; }
+  margin-bottom: 30px; }
 
 .margin-bottom-70 {
-  margin-bottom: 0px 0px 35px 0px; }
+  margin-bottom: 35px; }
 
 .margin-bottom-80 {
-  margin-bottom: 0px 0px 40px 0px; }
+  margin-bottom: 40px; }
 
 .margin-bottom-90 {
-  margin-bottom: 0px 0px 45px 0px; }
+  margin-bottom: 45px; }
 
 .margin-bottom-100 {
-  margin-bottom: 0px 0px 50px 0px; }
+  margin-bottom: 50px; }
 
 .margin-bottom-110 {
-  margin-bottom: 0px 0px 55px 0px; }
+  margin-bottom: 55px; }
 
 .margin-bottom-120 {
-  margin-bottom: 0px 0px 60px 0px; }
+  margin-bottom: 60px; }
 
 .margin-bottom-130 {
-  margin-bottom: 0px 0px 65px 0px; }
+  margin-bottom: 65px; }
 
 .margin-bottom-140 {
-  margin-bottom: 0px 0px 70px 0px; }
+  margin-bottom: 70px; }
 
 .margin-bottom-150 {
-  margin-bottom: 0px 0px 75px 0px; }
+  margin-bottom: 75px; }
 
 .margin-bottom-160 {
-  margin-bottom: 0px 0px 80px 0px; }
+  margin-bottom: 80px; }
 
 .margin-bottom-170 {
-  margin-bottom: 0px 0px 85px 0px; }
+  margin-bottom: 85px; }
 
 .margin-bottom-180 {
-  margin-bottom: 0px 0px 90px 0px; }
+  margin-bottom: 90px; }
 
 .margin-bottom-190 {
-  margin-bottom: 0px 0px 95px 0px; }
+  margin-bottom: 95px; }
 
 .margin-bottom-200 {
-  margin-bottom: 0px 0px 100px 0px; }
+  margin-bottom: 100px; }
 
 .margin-right-10 {
-  margin-right: 0px 0px 5px 0px; }
+  margin-right: 5px; }
 
 .margin-right-20 {
-  margin-right: 0px 0px 10px 0px; }
+  margin-right: 10px; }
 
 .margin-right-30 {
-  margin-right: 0px 0px 15px 0px; }
+  margin-right: 15px; }
 
 .margin-right-40 {
-  margin-right: 0px 0px 20px 0px; }
+  margin-right: 20px; }
 
 .margin-right-50 {
-  margin-right: 0px 0px 25px 0px; }
+  margin-right: 25px; }
 
 .margin-right-60 {
-  margin-right: 0px 0px 30px 0px; }
+  margin-right: 30px; }
 
 .margin-right-70 {
-  margin-right: 0px 0px 35px 0px; }
+  margin-right: 35px; }
 
 .margin-right-80 {
-  margin-right: 0px 0px 40px 0px; }
+  margin-right: 40px; }
 
 .margin-right-90 {
-  margin-right: 0px 0px 45px 0px; }
+  margin-right: 45px; }
 
 .margin-right-100 {
-  margin-right: 0px 0px 50px 0px; }
+  margin-right: 50px; }
 
 .margin-right-110 {
-  margin-right: 0px 0px 55px 0px; }
+  margin-right: 55px; }
 
 .margin-right-120 {
-  margin-right: 0px 0px 60px 0px; }
+  margin-right: 60px; }
 
 .margin-right-130 {
-  margin-right: 0px 0px 65px 0px; }
+  margin-right: 65px; }
 
 .margin-right-140 {
-  margin-right: 0px 0px 70px 0px; }
+  margin-right: 70px; }
 
 .margin-right-150 {
-  margin-right: 0px 0px 75px 0px; }
+  margin-right: 75px; }
 
 .margin-right-160 {
-  margin-right: 0px 0px 80px 0px; }
+  margin-right: 80px; }
 
 .margin-right-170 {
-  margin-right: 0px 0px 85px 0px; }
+  margin-right: 85px; }
 
 .margin-right-180 {
-  margin-right: 0px 0px 90px 0px; }
+  margin-right: 90px; }
 
 .margin-right-190 {
-  margin-right: 0px 0px 95px 0px; }
+  margin-right: 95px; }
 
 .margin-right-200 {
-  margin-right: 0px 0px 100px 0px; }
+  margin-right: 100px; }
 
 .margin-left-10 {
-  margin-left: 0px 0px 5px 0px; }
+  margin-left: 5px; }
 
 .margin-left-20 {
-  margin-left: 0px 0px 10px 0px; }
+  margin-left: 10px; }
 
 .margin-left-30 {
-  margin-left: 0px 0px 15px 0px; }
+  margin-left: 15px; }
 
 .margin-left-40 {
-  margin-left: 0px 0px 20px 0px; }
+  margin-left: 20px; }
 
 .margin-left-50 {
-  margin-left: 0px 0px 25px 0px; }
+  margin-left: 25px; }
 
 .margin-left-60 {
-  margin-left: 0px 0px 30px 0px; }
+  margin-left: 30px; }
 
 .margin-left-70 {
-  margin-left: 0px 0px 35px 0px; }
+  margin-left: 35px; }
 
 .margin-left-80 {
-  margin-left: 0px 0px 40px 0px; }
+  margin-left: 40px; }
 
 .margin-left-90 {
-  margin-left: 0px 0px 45px 0px; }
+  margin-left: 45px; }
 
 .margin-left-100 {
-  margin-left: 0px 0px 50px 0px; }
+  margin-left: 50px; }
 
 .margin-left-110 {
-  margin-left: 0px 0px 55px 0px; }
+  margin-left: 55px; }
 
 .margin-left-120 {
-  margin-left: 0px 0px 60px 0px; }
+  margin-left: 60px; }
 
 .margin-left-130 {
-  margin-left: 0px 0px 65px 0px; }
+  margin-left: 65px; }
 
 .margin-left-140 {
-  margin-left: 0px 0px 70px 0px; }
+  margin-left: 70px; }
 
 .margin-left-150 {
-  margin-left: 0px 0px 75px 0px; }
+  margin-left: 75px; }
 
 .margin-left-160 {
-  margin-left: 0px 0px 80px 0px; }
+  margin-left: 80px; }
 
 .margin-left-170 {
-  margin-left: 0px 0px 85px 0px; }
+  margin-left: 85px; }
 
 .margin-left-180 {
-  margin-left: 0px 0px 90px 0px; }
+  margin-left: 90px; }
 
 .margin-left-190 {
-  margin-left: 0px 0px 95px 0px; }
+  margin-left: 95px; }
 
 .margin-left-200 {
-  margin-left: 0px 0px 100px 0px; }
+  margin-left: 100px; }
 
 /* ------------------------------------------------------------------------
   - Media Size Variable
