+++ src/assets/scss/variable.scss
@@ -136,28 +136,28 @@ $margin: margin- !default;
 $margin-top: margin-top- !default;
 @for $i from 1 through 20 {
   .#{$margin-top}#{$i*10} {
-    margin-top: (5px * $i) 0px 0px 0px;
+    margin-top: (5px * $i);
   }
 }
 
 $margin-bottom: margin-bottom- !default;
 @for $i from 1 through 20 {
   .#{$margin-bottom}#{$i*10} {
-    margin-bottom: 0px 0px (5px * $i) 0px;
+    margin-bottom: (5px * $i)
   }
 }
 
 $margin-right: margin-right- !default;
 @for $i from 1 through 20 {
   .#{$margin-right}#{$i*10} {
-    margin-right: 0px 0px (5px * $i) 0px;
+    margin-right: (5px * $i);
   }
 }
 
 $margin-left: margin-left- !default;
 @for $i from 1 through 20 {
   .#{$margin-left}#{$i*10} {
-    margin-left: 0px 0px (5px * $i) 0px;
+    margin-left: (5px * $i);
   }
 }
