+++ src/components/items/notice/Notice.vue
@@ -1,31 +1,29 @@
 <template>
-  <div class="container">
-    <div class="row margin-20">
-      <div class="col-sm-4"
-      >
-
-      </div>
-    </div>
-  </div>
+  <el-carousel
+    :interval="5000"
+    arrow="always"
+    height="500px"
+  >
+    <el-carousel-item v-for="item in 4" :key="item">
+      <h3>{{ item }}</h3>
+    </el-carousel-item>
+  </el-carousel>
 </template>
 
-<script>
-  export default {
-    data () {
-      return {
-
-      }
-    },
-    methods: {
-    },
-    computed: {
-    },
-    watch: {
-    }
+<style>
+  .el-carousel__item h3 {
+    color: #475669;
+    font-size: 18px;
+    opacity: 0.75;
+    line-height: 300px;
+    margin: 0;
   }
-</script>
 
-<style lang="scss">
+  .el-carousel__item:nth-child(2n) {
+    background-color: #99a9bf;
+  }
 
+  .el-carousel__item:nth-child(2n+1) {
+    background-color: #d3dce6;
+  }
 </style>
-
